learnflow_api/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymysql==1.1.1
firebase-admin==6.5.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4